Add files via upload
This project uses a set of 50K IMBD Reviews that has been web scraped from IMDB.com to build a NLP machine learning model. The reviews are classified according to both positive and negative sentiments respectively. This dataset will be utilized to build a  classifier model that will be able to accurately predict both positive and negative sentiments.