Add files via upload
This project begins with web scraping the top 1000 movies from IMBD.com. From the list of top 1000 movies page we made separate requests to the respective movie pages to collect data like the budget, gross earnings, runtime and rating. After scraping the data this project continues with Exploratory Data Analysis which begins with cleaning the data and performing various analysis along with insights.